+++ src/modules/google-analytics/hooks/beforeRegistration.ts
@@ -2,7 +2,7 @@ import VueAnalytics from 'vue-analytics'
 import { router } from '@vue-storefront/core/app'
 import { Logger } from '@vue-storefront/core/lib/logger'
 
-export function beforeRegistration(Vue, config, store, isServer) {
+export function beforeRegistration({ Vue, config, store, isServer }) {
   if (config.analytics.id && !isServer) {
     Vue.use(VueAnalytics, {
       id: config.analytics.id,